backend_ddd/requirements.txt
Flask
psycopg2-binary
pytest
requests
gunicorn
python-dotenv
firebase-admin